@@ -16,10 +16,10 @@
     "interpolate-html-plugin": "^4.0.0",
     "json-schema-resolver": "1.3.0",
     "path": "^0.12.7",
-    "react": "=17.0.2",
+    "react": "^18.1.0",
     "react-app-polyfill": "^3.0.0",
     "react-dev-utils": "12.0.1",
-    "react-dom": "=17.0.2",
+    "react-dom": "^18.1.0",
     "react-hot-loader": "^4.13.0",
     "react-router-dom": "^6.3.0",
     "react-scripts": "^5.0.1",
